+++ packages/analytics-node/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.46](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-node@1.5.45...@amplitude/analytics-node@1.5.46) (2026-03-16)
+
+**Note:** Version bump only for package @amplitude/analytics-node
+
+
+
+
+
+## [1.5.45](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-node@1.5.44...@amplitude/analytics-node@1.5.45) (2026-03-13)
+
+**Note:** Version bump only for package @amplitude/analytics-node
+
+
+
+
+
+## [1.5.44](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-node@1.5.43...@amplitude/analytics-node@1.5.44) (2026-03-12)
+
+**Note:** Version bump only for package @amplitude/analytics-node
+
+
+
+
+
 ## [1.5.43](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-node@1.5.42...@amplitude/analytics-node@1.5.43) (2026-03-09)
 
 **Note:** Version bump only for package @amplitude/analytics-node